backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings

sqlmodel
pymysql
pillow
faiss-cpu
torch
clip-anytorch
python-jose[cryptography]
passlib[bcrypt]
python-multipart
transformers
requests

celery[redis]